requirements.txt
Django==1.11
Pillow==5.2.0
Unidecode==1.0.22
Willow==1.1
asn1crypto==0.23.0
beautifulsoup4==4.6.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
cryptography==2.0.2
django-cogwheels==0.2
django-comments-xtd==2.2.0
django-contrib-comments==1.9.0
django-cors-middleware==1.3.1
django-modelcluster==4.2
django-oauth-toolkit==1.0.0
django-taggit==0.23.0
django-tinymce==2.6.0
django-treebeard==4.3
django-widget-tweaks==1.4.3
djangorestframework==3.7.3
docutils==0.14
draftjs-exporter==2.1.2
html5lib==1.0.1
idna==2.6
oauthlib==2.0.3
psycopg2==2.7.3.2
pycparser==2.18
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
wagtail==2.1.3
wagtail-comments-xtd==0.2.9
wagtail-condensedinlinepanel==0.5.0
wagtailfontawesome==1.1.3
wagtailmenus==2.11
webencodings==0.5.1
